+++ src/clerk_backend_api/models/listorganizationmembershipsop.py
@@ -11,11 +11,11 @@ from typing_extensions import Annotated, NotRequired
 class ListOrganizationMembershipsRequestTypedDict(TypedDict):
     organization_id: str
     r"""The organization ID."""
-    limit: NotRequired[float]
+    limit: NotRequired[int]
     r"""Applies a limit to the number of results returned.
     Can be used for paginating the results together with `offset`.
     """
-    offset: NotRequired[float]
+    offset: NotRequired[int]
     r"""Skip the first `offset` results when paginating.
     Needs to be an integer greater or equal to zero.
     To be used in conjunction with `limit`.
@@ -30,11 +30,11 @@ class ListOrganizationMembershipsRequestTypedDict(TypedDict):
 class ListOrganizationMembershipsRequest(BaseModel):
     organization_id: Annotated[str, FieldMetadata(path=PathParamMetadata(style="simple", explode=False))]
     r"""The organization ID."""
-    limit: Annotated[Optional[float], FieldMetadata(query=QueryParamMetadata(style="form", explode=True))] = 10
+    limit: Annotated[Optional[int], FieldMetadata(query=QueryParamMetadata(style="form", explode=True))] = 10
     r"""Applies a limit to the number of results returned.
     Can be used for paginating the results together with `offset`.
     """
-    offset: Annotated[Optional[float], FieldMetadata(query=QueryParamMetadata(style="form", explode=True))] = 0
+    offset: Annotated[Optional[int], FieldMetadata(query=QueryParamMetadata(style="form", explode=True))] = 0
     r"""Skip the first `offset` results when paginating.
     Needs to be an integer greater or equal to zero.
     To be used in conjunction with `limit`.
